broker_setup/requirements.txt
confluent-kafka==2.2.0
http-status==1.0.0
six==1.16.0
python-dotenv==1.0.0